Create README.md
The code defines three variables: firstName, lastName, and age. The firstName variable is assigned the string value "Jawad", the lastName variable is assigned the string value "Ahmed", and the age variable is assigned the number value 28.

The code then uses a JavaScript template literal to create a string that displays a message, which is passed as an argument to the alert() function. The alert() function displays a message box to the user with the specified message.

The template literal evaluates the expressions within the ${} brackets and concatenates the results with the surrounding string. The resulting message is "Salam, Jawad Ahmed, your age is 28." which is displayed to the user through the alert() function.